Moved thread functions to threadPriorities file.
